[fix] props 가 너무 많아 context api를 이용해 리팩토링
