+++ src/components/TodoButton.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import styles from '../styles/todoButton.module.css';
 
 function TodoButton({ text, onClick }) {
+
     return (
-        <button className={styles.button} onClick={onClick}>
+        <button className={text === '등록' ? styles.registerButton : text === '삭제' ? styles.removeButton : styles.updateButton} onClick={onClick}>
             {text}
         </button>
     );
