+++ src/context/root.js
@@ -1,0 +1,6 @@
+import React from "react";
+
+const todoListContext = React.createContext();
+
+
+export default todoListContext;
